IMCMS-390 - Implement 2-factor authentication for ordinary login: - Remove unneeded unboxing

+++ server/src/com/imcode/imcms/api/TextDocument.java
@@ -1,8 +1,17 @@
 package com.imcode.imcms.api;
 
 import com.imcode.imcms.mapping.DocumentGetter;
-import imcode.server.document.*;
-import imcode.server.document.textdocument.*;
+import imcode.server.document.DirectDocumentReference;
+import imcode.server.document.DocumentDomainObject;
+import imcode.server.document.DocumentPredicate;
+import imcode.server.document.DocumentReference;
+import imcode.server.document.DocumentTypeDomainObject;
+import imcode.server.document.textdocument.ImageDomainObject;
+import imcode.server.document.textdocument.MenuDomainObject;
+import imcode.server.document.textdocument.MenuItemDomainObject;
+import imcode.server.document.textdocument.TextDocumentDomainObject;
+import imcode.server.document.textdocument.TextDomainObject;
+import imcode.server.document.textdocument.TreeSortKeyDomainObject;
 import imcode.server.user.UserDomainObject;
 import org.apache.commons.collections.CollectionUtils;
 import org.apache.commons.collections.Predicate;
@@ -11,7 +20,12 @@ import org.apache.commons.collections.functors.CloneTransformer;
 import org.apache.commons.collections.map.TransformedSortedMap;
 import org.apache.commons.lang.StringUtils;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.List;
+import java.util.Map;
+import java.util.SortedMap;
+import java.util.TreeMap;
 
 /**
  * Document that can contain texts, images, menus, includes.
@@ -241,7 +255,7 @@ public class TextDocument extends Document {
     public Document getInclude(int includeIndexInDocument) {
         Integer includedDocumentId = getInternalTextDocument().getIncludedDocumentId(includeIndexInDocument);
         if (null != includedDocumentId) {
-            DocumentDomainObject includedDocument = getDocumentGetter().getDocument(includedDocumentId.intValue());
+            DocumentDomainObject includedDocument = getDocumentGetter().getDocument(includedDocumentId);
             if (null != includedDocument) {
                 return DocumentService.wrapDocumentDomainObject(includedDocument, contentManagementSystem);
             }
